dbeaver/dbeaver-ee#1040 move UI method from model to UI class, change json file constant for task folders
