+++ plugins/org.jkiss.dbeaver.model/src/org/jkiss/dbeaver/model/task/DBTTaskManager.java
@@ -79,10 +79,6 @@ public interface DBTTaskManager {
 
     void removeTaskFolder(@NotNull DBTTaskFolder taskFolder) throws DBException;
 
-    DBTTaskFolder getCurrentSelectedTaskFolder();
-
-    void setCurrentSelectedTaskFolder(DBTTaskFolder taskFolder);
-
     @NotNull
     File getStatisticsFolder();
