+++ plugins/org.jkiss.dbeaver.tasks.ui.view/src/org/jkiss/dbeaver/tasks/ui/view/TaskHandlerCreate.java
@@ -18,34 +18,33 @@ package org.jkiss.dbeaver.tasks.ui.view;
 
 import org.eclipse.core.commands.AbstractHandler;
 import org.eclipse.core.commands.ExecutionEvent;
-import org.eclipse.core.commands.ExecutionException;
 import org.eclipse.jface.dialogs.IDialogConstants;
 import org.eclipse.jface.viewers.ISelection;
 import org.eclipse.jface.viewers.IStructuredSelection;
 import org.eclipse.ui.handlers.HandlerUtil;
-import org.jkiss.dbeaver.model.app.DBPProject;
 import org.jkiss.dbeaver.model.task.DBTTaskFolder;
-import org.jkiss.dbeaver.model.task.DBTTaskManager;
 import org.jkiss.dbeaver.runtime.DBWorkbench;
+import org.jkiss.dbeaver.tasks.ui.wizard.TaskConfigurationWizard;
 import org.jkiss.dbeaver.tasks.ui.wizard.TaskConfigurationWizardDialog;
-import org.jkiss.dbeaver.ui.navigator.NavigatorUtils;
 
 public class TaskHandlerCreate extends AbstractHandler {
 
     @Override
-    public Object execute(ExecutionEvent event) throws ExecutionException {
-        DBPProject project = NavigatorUtils.getSelectedProject();
+    public Object execute(ExecutionEvent event) {
 
         TaskConfigurationWizardDialog dialog = new TaskConfigurationWizardDialog(HandlerUtil.getActiveWorkbenchWindow(event));
-        DBTTaskManager taskManager = project.getTaskManager();
-        final ISelection selection = HandlerUtil.getCurrentSelection(event);
-        if (selection instanceof IStructuredSelection) {
-            IStructuredSelection structuredSelection = (IStructuredSelection) selection;
-            if (structuredSelection.size() == 1 && structuredSelection.getFirstElement() instanceof DBTTaskFolder) {
-                taskManager.setCurrentSelectedTaskFolder((DBTTaskFolder) structuredSelection.getFirstElement());
+        TaskConfigurationWizard taskWizard = dialog.getTaskWizard();
+        if (taskWizard != null) {
+            final ISelection selection = HandlerUtil.getCurrentSelection(event);
+            if (selection instanceof IStructuredSelection) {
+                IStructuredSelection structuredSelection = (IStructuredSelection) selection;
+                if (structuredSelection.size() == 1 && structuredSelection.getFirstElement() instanceof DBTTaskFolder) {
+                    // We need to throw this information about the choice of task folder in TaskConfigurationWizardPageTask
+                    taskWizard.setCurrentSelectedTaskFolder((DBTTaskFolder) structuredSelection.getFirstElement());
+                }
+            } else {
+                taskWizard.setCurrentSelectedTaskFolder(null);
             }
-        } else {
-            taskManager.setCurrentSelectedTaskFolder(null);
         }
         if (dialog.open() == IDialogConstants.OK_ID) {
