+++ plugins/org.jkiss.dbeaver.tasks.ui/src/org/jkiss/dbeaver/tasks/ui/wizard/TaskConfigurationWizardDialog.java
@@ -92,7 +92,7 @@ public class TaskConfigurationWizardDialog extends ActiveWizardDialog {
         return (TaskConfigurationWizard) super.getWizard();
     }
 
-    protected TaskConfigurationWizard getTaskWizard() {
+    public TaskConfigurationWizard getTaskWizard() {
         return (TaskConfigurationWizard) super.getWizard();
     }
